Organise code better
