+++ src/commands/change-case-opcode.ts
@@ -1,9 +1,9 @@
 import * as vscode from 'vscode';
-import { TokenKind } from './common';
-import { OpcodeCase, WorkspaceManager } from './managers';
-import { AssemblyLine } from './parsers';
-import { convertToCase } from './utilities';
-import { Logger } from './logger';
+import { convertToCase } from '../common';
+import { OpcodeCase, TokenKind } from '../constants';
+import { WorkspaceManager } from '../managers';
+import { AssemblyLine } from '../parsers';
+import { Logger } from '../logger';
 
 export class ChangeCaseOpcodeCommand {
