+++ src/common/assembly-file-reference.ts
@@ -1,0 +1,10 @@
+import * as vscode from 'vscode';
+
+export class AssemblyFileReference {
+  constructor(
+    public uri: vscode.Uri,
+    public range: vscode.Range,
+  ) {
+  }
+}
+
