+++ src/constants/configuration.ts
@@ -1,0 +1,23 @@
+export enum Command {
+  lwasm,
+  xroar
+}
+
+export enum OSPlatform {
+  windows,
+  macOS,
+  linux
+}
+
+export enum OpcodeCase {
+  lowercase = 'lowercase',
+  uppercase = 'uppercase',
+  capitalised = 'capitalised',
+}
+
+export enum HelpVerbosity {
+  none = 'none',
+  light = 'light',
+  full = 'full',
+}
+
