+++ src/providers/workspace-symbol.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
+import { convertToSymbolKind } from '../common';
 import { WorkspaceManager } from '../managers';
-import { convertToSymbolKind } from '../utilities';
 
 export class WorkspaceSymbolProvider implements vscode.WorkspaceSymbolProvider {
